code editor now hasvisible content while loading monaco
