+++ elements/hax-body/lib/hax-ce-context.js
@@ -1,8 +1,8 @@
 import { LitElement, html, css } from "lit-element/lit-element.js";
 import { HAXStore } from "./hax-store.js";
 import "@lrnwebcomponents/hax-body/lib/hax-context-item.js";
-import "@lrnwebcomponents/hax-body/lib/hax-button-menu.js";
-import "@lrnwebcomponents/hax-body/lib/hax-button-menu-item.js";
+import "@lrnwebcomponents/hax-body/lib/hax-toolbar-menu.js";
+import "@lrnwebcomponents/hax-body/lib/hax-toolbar-menu-item.js";
 import { HaxToolbarBehaviors } from "@lrnwebcomponents/hax-body/lib/hax-toolbar.js";
 import { wipeSlot } from "@lrnwebcomponents/utils/utils";
 /**
@@ -67,8 +67,8 @@ class HaxCeContext extends HaxToolbarBehaviors(LitElement) {
           ?disabled="${!this.sourceView}"
           event-name="hax-source-view-toggle"
         ></hax-context-item>
-        <hax-button-menu icon="add" label="Insert item above or below">
-          <hax-button-menu-item slot="menuitem">
+        <hax-toolbar-menu icon="add" label="Insert item above or below">
+          <hax-toolbar-menu-item slot="menuitem">
             <hax-context-item
               action
               show-text-label
@@ -77,8 +77,8 @@ class HaxCeContext extends HaxToolbarBehaviors(LitElement) {
               event-name="insert-above-active"
               label="Insert item above"
             ></hax-context-item>
-          </hax-button-menu-item>
-          <hax-button-menu-item slot="menuitem">
+          </hax-toolbar-menu-item>
+          <hax-toolbar-menu-item slot="menuitem">
             <hax-context-item
               action
               show-text-label
@@ -87,8 +87,8 @@ class HaxCeContext extends HaxToolbarBehaviors(LitElement) {
               event-name="insert-below-active"
               label="Insert item below"
             ></hax-context-item>
-          </hax-button-menu-item>
-        </hax-button-menu>
+          </hax-toolbar-menu-item>
+        </hax-toolbar-menu>
         <slot name="secondary"></slot>
         <slot name="more"></slot>
       </div>
