+++ elements/hax-body/package.json
@@ -37,6 +37,7 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@lrnwebcomponents/a11y-collapse": "^3.0.6",
+    "@lrnwebcomponents/a11y-menu-button": "^3.0.6",
     "@lrnwebcomponents/code-editor": "^3.0.6",
     "@lrnwebcomponents/code-sample": "^3.0.6",
     "@lrnwebcomponents/dl-behavior": "^3.0.6",
@@ -55,6 +56,7 @@
     "@lrnwebcomponents/simple-icon-picker": "^3.0.6",
     "@lrnwebcomponents/simple-modal": "^3.0.6",
     "@lrnwebcomponents/simple-popover": "^3.0.6",
+    "@lrnwebcomponents/simple-toolbar": "^3.0.6",
     "@lrnwebcomponents/simple-toast": "^3.0.6",
     "@lrnwebcomponents/simple-tooltip": "^3.0.5",
     "@lrnwebcomponents/undo-manager": "^3.0.5",
